Restructure testserver
