+++ scripts/packages/VSCodeTestServer/src/VSCodeTestServer.jl
@@ -1,167 +1,15 @@
 module VSCodeTestServer
 
-include("../../URIParser/src/URIParser.jl")
-include("../../JSON/src/JSON.jl")
-include("../../OrderedCollections/src/OrderedCollections.jl")
-include("../../CodeTracking/src/CodeTracking.jl")
-
-module JSONRPC
-import ..JSON
-import UUIDs
-include("../../JSONRPC/src/packagedef.jl")
-end
-
-module JuliaInterpreter
-using ..CodeTracking
-include("../../JuliaInterpreter/src/packagedef.jl")
-end
-
-module LoweredCodeUtils
-using ..JuliaInterpreter
-using ..JuliaInterpreter: SSAValue, SlotNumber, Frame
-using ..JuliaInterpreter: @lookup, moduleof, pc_expr, step_expr!, is_global_ref, is_quotenode_egal, whichtt,
-    next_until!, finish_and_return!, get_return, nstatements, codelocation, linetable,
-    is_return, lookup_return, is_GotoIfNot, is_ReturnNode
-
-include("../../LoweredCodeUtils/src/packagedef.jl")
-end
-
-module Revise
-using ..OrderedCollections
-using ..LoweredCodeUtils
-using ..CodeTracking
-using ..JuliaInterpreter
-using ..CodeTracking: PkgFiles, basedir, srcfiles, line_is_decl, basepath
-using ..JuliaInterpreter: whichtt, is_doc_expr, step_expr!, finish_and_return!, get_return,
-    @lookup, moduleof, scopeof, pc_expr, is_quotenode_egal,
-    linetable, codelocs, LineTypes, is_GotoIfNot, isassign, isidentical
-using ..LoweredCodeUtils: next_or_nothing!, trackedheads, callee_matches
-include("../../Revise/src/packagedef.jl")
-end
-
-# module DebugAdapter
-# import ..JuliaInterpreter
-# import ..JSON
-# import ..JSONRPC
-# import ..JSONRPC: @dict_readable, Outbound
-
-# include("../../DebugAdapter/src/packagedef.jl")
-# end
+include("pkg_imports.jl")
 
 import .JSONRPC: @dict_readable
 import Test
 
-struct VSCodeTestSet <: Test.AbstractTestSet
-    description::AbstractString
-    results::Vector
-    children::Vector
-    VSCodeTestSet(desc) = new(desc, [], [])
-end
+include("testserver_protocol.jl")
+include("helper.jl")
+include("vscode_testset.jl")
 
 const conn_endpoint = Ref{Union{Nothing,JSONRPC.JSONRPCEndpoint}}(nothing)
-
-@dict_readable struct Position
-    line::Int
-    character::Int
-end
-
-struct Range
-    start::Position
-    stop::Position
-end
-function JSON.lower(a::Range)
-    Dict("start" => a.start, "end" => a.stop)
-end
-
-@dict_readable struct Location
-    uri::String
-    range::Range
-end
-
-@dict_readable struct TestMessage
-    message::String
-    # expectedOutput?: string;
-    # actualOutput?: string;
-    location::Union{Nothing,Location}
-end
-
-JSONRPC.@dict_readable struct TestserverRunTestitemRequestParams <: JSONRPC.Outbound
-    uri::String
-    line::Int
-    column::Int
-    code::String
-end
-
-JSONRPC.@dict_readable struct TestserverRunTestitemRequestParamsReturn <: JSONRPC.Outbound
-    status::String
-    message::Union{Vector{TestMessage},Nothing}
-end
-
-const testserver_revise_request_type = JSONRPC.RequestType("testserver/revise", Nothing, String)
-const testserver_run_testitem_request_type = JSONRPC.RequestType("testserver/runtestitem", TestserverRunTestitemRequestParams, TestserverRunTestitemRequestParamsReturn)
-
-# TODO Use our new Uri2 once it is ready
-function uri2filepath(uri::AbstractString)
-    parsed_uri = try
-        URIParser.URI(uri)
-    catch
-        throw(LSUriConversionFailure("Cannot parse `$uri`."))
-    end
-
-    if parsed_uri.scheme !== "file"
-        return nothing
-    end
-
-    path_unescaped = URIParser.unescape(parsed_uri.path)
-    host_unescaped = URIParser.unescape(parsed_uri.host)
-
-    value = ""
-
-    if host_unescaped != "" && length(path_unescaped) > 1
-        # unc path: file://shares/c$/far/boo
-        value = "//$host_unescaped$path_unescaped"
-    elseif length(path_unescaped) >= 3 &&
-           path_unescaped[1] == '/' &&
-           isascii(path_unescaped[2]) && isletter(path_unescaped[2]) &&
-           path_unescaped[3] == ':'
-        # windows drive letter: file:///c:/far/boo
-        value = lowercase(path_unescaped[2]) * path_unescaped[3:end]
-    else
-        # other path
-        value = path_unescaped
-    end
-
-    if Sys.iswindows()
-        value = replace(value, '/' => '\\')
-    end
-
-    value = normpath(value)
-
-    return value
-end
-
-# TODO Use our new Uri2 once it is ready
-function filepath2uri(file::String)
-    isabspath(file) || error("Relative path `$file` is not valid.")
-    if Sys.iswindows()
-        file = normpath(file)
-        file = replace(file, "\\" => "/")
-        file = URIParser.escape(file)
-        file = replace(file, "%2F" => "/")
-        if startswith(file, "//")
-            # UNC path \\foo\bar\foobar
-            return string("file://", file[3:end])
-        else
-            # windows drive letter path
-            return string("file:///", file)
-        end
-    else
-        file = normpath(file)
-        file = URIParser.escape(file)
-        file = replace(file, "%2F" => "/")
-        return string("file://", file)
-    end
-end
 
 function run_revise_handler(conn, params::Nothing)
     try
@@ -192,7 +40,7 @@ function run_testitem_handler(conn, params::TestserverRunTestitemRequestParams)
     code_without_begin_end = params.code[6:end-3]
     code = string('\n'^params.line, ' '^params.column, code_without_begin_end)
 
-    ts = VSCodeTestSet("WE NEED A DESCRIPTION")
+    ts = VSCodeTestSet("")
 
     Test.push_testset(ts)
 
@@ -237,23 +85,6 @@ function run_testitem_handler(conn, params::TestserverRunTestitemRequestParams)
             "failed",
             [TestMessage(sprint(Base.show, i), Location(filepath2uri(string(i.source.file)), Range(Position(i.source.line - 1, 0), Position(i.source.line - 1, 0)))) for i in failed_tests]
         )
-    end
-end
-
-function Test.record(ts::VSCodeTestSet, res)
-    push!(ts.results, res)
-end
-
-function Test.record(ts::VSCodeTestSet, res::VSCodeTestSet)
-    push!(ts.children, res)
-end
-
-function Test.finish(ts::VSCodeTestSet)
-    if Test.get_testset_depth() != 0
-        # Attach this test set to the parent test set
-        parent_ts = Test.get_testset()
-        Test.record(parent_ts, ts)
-        return ts
     end
 end
