+++ scripts/packages/VSCodeTestServer/src/helper.jl
@@ -1,0 +1,62 @@
+# TODO Use our new Uri2 once it is ready
+function uri2filepath(uri::AbstractString)
+    parsed_uri = try
+        URIParser.URI(uri)
+    catch
+        throw(LSUriConversionFailure("Cannot parse `$uri`."))
+    end
+
+    if parsed_uri.scheme !== "file"
+        return nothing
+    end
+
+    path_unescaped = URIParser.unescape(parsed_uri.path)
+    host_unescaped = URIParser.unescape(parsed_uri.host)
+
+    value = ""
+
+    if host_unescaped != "" && length(path_unescaped) > 1
+        # unc path: file://shares/c$/far/boo
+        value = "//$host_unescaped$path_unescaped"
+    elseif length(path_unescaped) >= 3 &&
+           path_unescaped[1] == '/' &&
+           isascii(path_unescaped[2]) && isletter(path_unescaped[2]) &&
+           path_unescaped[3] == ':'
+        # windows drive letter: file:///c:/far/boo
+        value = lowercase(path_unescaped[2]) * path_unescaped[3:end]
+    else
+        # other path
+        value = path_unescaped
+    end
+
+    if Sys.iswindows()
+        value = replace(value, '/' => '\\')
+    end
+
+    value = normpath(value)
+
+    return value
+end
+
+# TODO Use our new Uri2 once it is ready
+function filepath2uri(file::String)
+    isabspath(file) || error("Relative path `$file` is not valid.")
+    if Sys.iswindows()
+        file = normpath(file)
+        file = replace(file, "\\" => "/")
+        file = URIParser.escape(file)
+        file = replace(file, "%2F" => "/")
+        if startswith(file, "//")
+            # UNC path \\foo\bar\foobar
+            return string("file://", file[3:end])
+        else
+            # windows drive letter path
+            return string("file:///", file)
+        end
+    else
+        file = normpath(file)
+        file = URIParser.escape(file)
+        file = replace(file, "%2F" => "/")
+        return string("file://", file)
+    end
+end
